Changed the majority of integer variables to regular ints.
The narrowing conversion errors and code riddled with static casts. It
just wasn't worth it.

+++ src/server/Ball.h
@@ -26,11 +26,11 @@ namespace pong
   struct Ball
   {
     explicit Ball(BallID id = 0,
-                  math::vector<int16_t> pos = math::vector<int16_t>(),
-                  int16_t diameter = 0)
+                  math::vector<int> pos = math::vector<int>(),
+                  int diameter = 0)
              : id(id), pos(pos), diameter(diameter){}
     BallID id;
-    math::vector<int16_t> pos;
-    int16_t diameter;
+    math::vector<int> pos;
+    int diameter;
   };
 }
