+++ src/server/Paddle.h
@@ -32,12 +32,12 @@ namespace pong
   struct Paddle
   {
     explicit Paddle(PaddleID id = 0,
-                    math::vector<int16_t> pos =math::vector<int16_t>(),
-                    int16_t width = 0,
-                    int16_t height = 0)
+                    math::vector<int> pos =math::vector<int>(),
+                    int width = 0,
+                    int height = 0)
              : id(id), pos(pos), width(width), height(height){}
     PaddleID id;
-    math::vector<int16_t> pos;
-    int16_t width, height;
+    math::vector<int> pos;
+    int width, height;
   };
 }
